feat: simplify the cucumber parser
This change simplifies the cucumber parsers and brings them down to 1 parser.
The docs are also updated with usage examples.

Closes #71

BREAKING CHANGE: cucumberMulti parser can't be used anymore

To migrate change `cucumberMulti` to `cucumber`
