+++ docs/cucumber.md
@@ -1,32 +1,105 @@
 # Cucumberjs parser
-There are 2 parsers that can be used, the `cucumber`- and the `cucumberMulti`-parser. Just select one of them based on the below description and use them like this
 
-```
+## Intro
+There are multiple version of CucumberJS that all differ a bit in the way they log failures. This parser will parse the log based on 2 log lines. That combination will determine if a `*.feature`-file needs to be run again.
+The 2 log lines are
+
+- `Failures:`
+- `Specs: /path/to/your/featurefile/flaky.feature`.
+
+The `Failures` log line is always printed on failed scenario's. The `Specs: /path/to/your/featurefile/flaky.feature` is not always printed, due to the capability settings in the Protractor configuration. See [here](#capability-configuration)
+
+
+## Usage
+The following 2 things need to be done to get the parser working
+
+### Configure protractor-flake
+Via the CLI:
+```shell
 // how to use the `cucumber`-parser
 protractor-flake --protractor-path=/path/to/protractor --parser cucumber --node-bin node --max-attempts=3 -- path/to/protractor.conf.js
-
-// how to use the `cucumberMulti`-parser
-protractor-flake --protractor-path=/path/to/protractor --parser cucumberMulti --node-bin node --max-attempts=3 -- path/to/protractor.conf.js
 ```
 
+Or programmatically:
 
+```js
+var protractorFlake = require('protractor-flake');
 
-## cucumber
-> This parser only works with `cucumberjs` version < v0.9.0
+// Default Options
+protractorFlake({
+  parser: 'cucumber'
+}, function (status, output) {
+  proces.exit(status)
+})
+```
 
-The `cucumber` parser will search the testoutput for this piece of text `/path/to/your/featurefile/flakey.feature:4 # Scenario: Flakey scenario`. This is the line that says that a scenario in a featurefile failed.
-The `/path/to/your/featurefile/flakey.feature` piece will be stripped out and passed to `protractor-flake` to rerun the file(s) for the amount of retries that have been given.
+### Add a `Specs: ..` log statement
+Depending on the version of CucumberJS that is used 1 of the hooks defined below can be used. This can be saved to a file like for example `afterHooks.js`. Make sure that this file is required in the `protractor.conf.js` file like this
 
-## cucumberMulti
-> This parser is based on `cucumberjs` > v0.9.0. These versions have a different logging and also a different way in returning the path of the runned specs. (See example [logging](../test/unit/support/fixtures/cucumberjs/)).
+```js
+cucumberOpts: {
+    require: [
+        'path/to/your/afterHooks.js'
+    ]
+}
+```
 
-> **This parser can parse testoutput from `capabilities` and `multiCapabilities`**
+This will **always** print the `Specs:` in the testoutput when a scenario fails. The parser will filter double specs.
 
-The `cucumberMulti` parser will search the testoutput for the text combination `Failures:`and `Specs: /path/to/your/featurefile/flakey.feature`. 
-The parser needs to search for both pieces of text because succeeded specs are also being logged.
-The `/path/to/your/featurefile/flakey.feature` piece will be stripped out and passed to `protractor-flake` to rerun the file(s) for the amount of retries that have been given.
+#### Version 1
+```js
+'use strict';
 
-Based on the way the capability in Protractor has been configured the `Specs:` are printed. See the table below for the results.	
+module.exports = function () {
+    this.After(function (scenarioResult) {
+        if (scenarioResult.isFailed()) {
+            // Log the spec to the console for protractor-flake to be able to rerun the failed specs
+            console.log('Specs:', scenarioResult.getUri());
+        }
+
+        return Promise.resolve();
+    });
+};
+```
+
+#### Version 2
+```js
+'use strict';
+
+const {defineSupportCode} = require('cucumber');
+
+defineSupportCode(({After}) => {
+    After(function (scenarioResult) {
+        if (scenarioResult.status === 'failed') {
+            // Log the spec to the console for protractor-flake to be able to rerun the failed specs
+            console.log('Specs:', scenarioResult.scenario.uri);
+        }
+
+        return Promise.resolve();
+    });
+});
+```
+
+#### Version 3
+```js
+'use strict';
+
+const {defineSupportCode, Status} = require('cucumber');
+
+defineSupportCode(({After}) => {
+    After(function (testCase) {
+        if (testCase.status === 'failed') {
+            // Log the spec to the console for protractor-flake to be able to rerun the failed specs
+            console.log('Specs:', testCase.scenario.uri);
+        }
+
+        return Promise.resolve();
+    });
+});
+```
+
+## Capability configuration
+Based on the way the capability in Protractor has been configured the `Specs:` are printed. See the table below for the results.
 
 ```
 // Single
@@ -65,46 +138,3 @@ multiCapabilities = [
 |  multi     |     false      |   >1    |   true   |   **FALSE**   |
 |  multi     |      true      |   >1    |  false   |    **TRUE**   |
 |  multi     |     false      |   >1    |  false   |     false     |
-
-When no `Specs:` are printed in the logging Protractor will run the Protractor command again which means that all the specs are run on **ALL** the instances (also the succeeded specs).
-
-### Keep in mind
-Keep in mind that if the testsuite holds multiple featurefiles and only 1 featurefile is flakey the following will occur. For example:
-* Run 1:
-  * 2 featurefiles are run 
-  * 1 is flakey => `protractor-flake` will parse the output, finds 1 failed specs output (see table) and reruns protractor
-* Run 2:
-  * 1 featurefile is run
-  * featurefile is flakey => `protractor-flake` will parse the output and finds NO specs in the output (see table) and reruns protractor
-* Run 3:
-  * protractor is rerun by `protractor-flake` but gets no `--specs`, this means that ALL the featurefiles are run again 
-  
-To avoid this behaviour a simple "workaround" can be implemented, see "Always print specs"
-
-### Always print specs
-There is a way to always print `Specs:`. This can be done with the following hook, see [cucumberjs](https://github.com/cucumber/cucumber-js) for more hook info:
-
-```
-var afterHook = function () {
-  this.After(function (scenario, callback) {
-        if (scenario.isFailed()) {
-            console.log('Specs:', scenario.getUri());
-        }
-        callback();
-  });
-};
-
-module.exports = afterHook;
-```
-
-Make sure that the file is required in the `protractor.conf.js` file like this
-
-```
-cucumberOpts: {
-    require: [
-        'path/to/your/afterHooks.js'
-    ]
-}
-```
-
-This will always print the `Specs:` in the testoutput. The parser will filter double specs.
