+++ test/unit/parsers/cucumber.test.js
@@ -2,19 +2,60 @@ import readFixture from '../support/read-fixture'
 import cucumberParser from '../../../src/parsers/cucumber'
 
 describe('cucumberParser', () => {
-  const failedOutput = readFixture('failed-cucumberjs-output.txt')
-
   describe('#parse', () => {
-    it('properly identifies failed cucumberjs feature files', () => {
-      expect(cucumberParser.parse(failedOutput)).to.eql([
-        '/Users/jrust/code/features/automated/fail.feature'
+    it('properly handles error output in multicapabilities tests with a feature', function () {
+      let output = readFixture('cucumberjs/cucumberjs-multi-output-feature-failure.txt')
+
+      expect(cucumberParser.parse(output)).to.eql([
+        '/Users/wswebcreation/test/e2e/features/functional/flakey.feature'
       ])
     })
 
-    it('returns an empty array if cucumberjs output has no matches', () => {
-      let successOutput = readFixture('success-cucumberjs-output.txt')
+    it('properly handles error output in sharded multicapabilities tests with a feature', function () {
+      let output = readFixture('cucumberjs/cucumberjs-multi-output-shared-feature-failure.txt')
 
-      expect(cucumberParser.parse(successOutput)).to.eql([])
+      expect(cucumberParser.parse(output)).to.eql([
+        '/Users/wswebcreation/test/e2e/features/functional/flakey.feature'
+      ])
+    })
+
+    it('returns an empty array if cucumberjs output has no matches for a multicapabilities tests with a feature', function () {
+      let output = readFixture('cucumberjs/cucumberjs-multi-output-feature-success.txt')
+
+      expect(cucumberParser.parse(output)).to.eql([])
+    })
+
+    it('returns an empty array if cucumberjs output has no matches for a sharded multicapabilities tests with a feature', function () {
+      let output = readFixture('cucumberjs/cucumberjs-multi-output-shared-feature-success.txt')
+
+      expect(cucumberParser.parse(output)).to.eql([])
+    })
+
+    it('can\'t handles error output in multicapabilities tests with features', function () {
+      let output = readFixture('cucumberjs/cucumberjs-multi-output-features-failures.txt')
+
+      expect(cucumberParser.parse(output)).to.eql([])
+    })
+
+    it('properly handles error output in sharded multicapabilities tests with features and double failures', function () {
+      let output = readFixture('cucumberjs/cucumberjs-multi-output-shared-features-failures.txt')
+
+      expect(cucumberParser.parse(output)).to.eql([
+        '/Users/wswebcreation/protractor-flake-tests/features/flakey.feature',
+        '/Users/wswebcreation/protractor-flake-tests/features/another.flakey.feature'
+      ])
+    })
+
+    it('returns an empty array if cucumberjs output has no matches for a multicapabilities tests with features', function () {
+      let output = readFixture('cucumberjs/cucumberjs-multi-output-features-success.txt')
+
+      expect(cucumberParser.parse(output)).to.eql([])
+    })
+
+    it('returns an empty array if cucumberjs output has no matches for a sharded multicapabilities tests with features', function () {
+      let output = readFixture('cucumberjs/cucumberjs-multi-output-shared-features-success.txt')
+
+      expect(cucumberParser.parse(output)).to.eql([])
     })
   })
 })
